fix: fold NOT_COMPARABLE scan-set members into the set's exit code
`_aggregate_scan_set_verdict` had no NOT_COMPARABLE handling: a set with
one NOT_COMPARABLE member and otherwise-compatible members previously
exited 0, even though `ScanSetResult.to_dict()`'s `run_outcome.operational`
already named the refusal via `member_not_comparable`. Since the
`--artifact-set` CLI exits solely off `result.exit_code`, this silently
reported success on a set CI should have treated as blocked.

A NOT_COMPARABLE member now becomes the reported verdict/exit 6 (ADR-050
D2's convention) whenever no stronger API_BREAK/BREAKING already won --
outranking a sibling EVIDENCE_CONTRACT_ERROR there too, since "couldn't
compare at all" is a stronger operational signal -- but a real break from
a different member is never masked.

New tests in tests/test_scan_writers_run_outcome.py (moved there from
tests/test_scan_artifact_set.py, which sits at its own debt.yaml
no_growth baseline) verified via git-stash against the pre-fix commit to
fail pre-fix, pass post-fix.

Codex review (P1), fresh evidence on PR #1003 commit 0a8a0145.
